Share HTML parser tree fixture helpers
