+++ code/packages/rust/html-parser/CHANGELOG.md
@@ -11,6 +11,8 @@ documented in this file.
 - Generated WHATWG tree-insertion audit fixture over the html5lib
   adoption-agency, table/foster-parenting, template, foreign-content fragment,
   and HTML fragment-shell families, with a dedicated parser regression test.
+- Shared html5lib tree-construction test helpers keep smoke and focused
+  tree-insertion audit checks on the same parser and DOM dump path.
 - Stack-of-open-elements tree construction seed with void element handling,
   adjacent text merging, simple implied end tags, and unmatched end-tag
   diagnostics.
